+++ smartcontract/programs/doublezero-serviceability/tests/link_wan_test.rs
@@ -1013,7 +1013,7 @@ async fn test_wan_link_rejects_cyoa_interface() {
                 loopback_type: LoopbackType::None,
                 interface_cyoa: InterfaceCYOA::GREOverDIA,
                 bandwidth: 0,
-                ip_net: None,
+                ip_net: Some("100.1.0.0/31".parse().unwrap()),
                 cir: 0,
                 mtu: 1500,
                 routing_mode: RoutingMode::Static,